4.21: Part 4 - Exercise 21

+++ part4/bloglist/controllers/blogs.js
@@ -39,8 +39,28 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog) {
+        return response.status(404).json({ error: 'blog not found' })
+    }
+    else if (user.id === blog.user.toString()) {
+        await Blog.findByIdAndRemove(request.params.id)
+
+        user.blogs = user.blogs.filter(blog => blog._id.toString() !== request.params.id )
+        await user.save()
+
+        response.status(204).end()
+    }
+    else {
+        return response.status(401).json({ error: 'user not authorized to delete' })
+    }
 })
 
 blogsRouter.put('/:id', (request, response, next) => {
